pretixprint/requirements.txt
requests
brother_ql
urllib3
Pillow
python_dateutil
pyserial
pyusb
nameparser
